+++ tidy3d/components/subpixel_spec.py
@@ -39,7 +39,8 @@ class Staircasing(AbstractSubpixelAveragingMethod):
 
 
 class PolarizedAveraging(AbstractSubpixelAveragingMethod):
-    """Apply a polarized subpixel averaging method to dielectric boundaries.
+    """Apply a polarized subpixel averaging method to dielectric boundaries, which
+    is a phenomenological approximation of :class:`.ContourPathAveraging`.
 
     Note
     ----
@@ -51,7 +52,20 @@ class PolarizedAveraging(AbstractSubpixelAveragingMethod):
     """
 
 
-DielectricSubpixelType = Union[Staircasing, PolarizedAveraging]
+class ContourPathAveraging(AbstractSubpixelAveragingMethod):
+    """Apply a contour-path subpixel averaging method to dielectric boundaries.
+
+    Note
+    ----
+    The algorithm is based on:
+
+        A. Mohammadi, H. Nadgaran and M. Agio, "Contour-path effective
+        permittivities for the two-dimensional finite-difference
+        time-domain method", Optics express, 13(25), 10367-10381 (2005).
+    """
+
+
+DielectricSubpixelType = Union[Staircasing, PolarizedAveraging, ContourPathAveraging]
 
 
 class VolumetricAveraging(AbstractSubpixelAveragingMethod):